State writing concurrency improvement (upstream issue: commit blocks to state using a separate task #4937), added HeightDiff and height ops fixed, several read requests forwarded to ReadStateService

+++ zebra-consensus/src/checkpoint/list/tests.rs
@@ -292,7 +292,7 @@ fn checkpoint_list_hard_coded_max_gap(network: Network) -> Result<(), BoxError> 
     assert_eq!(heights.next(), Some(&previous_height));
 
     for height in heights {
-        let height_limit = (previous_height + (crate::MAX_CHECKPOINT_HEIGHT_GAP as i32)).unwrap();
+        let height_limit = (previous_height + (crate::MAX_CHECKPOINT_HEIGHT_GAP as i64)).unwrap();
         assert!(
             height <= &height_limit,
             "Checkpoint gaps must be within MAX_CHECKPOINT_HEIGHT_GAP"
